fix: cibuildwheel local build — two fixes for source-from-scratch build
1. dnf: add vim-common so xxd is available for embedding .cl kernels
   Qrack's cmake uses xxd to convert OpenCL kernel source files into
   C++ headers.  The manylinux_2_34 container has cmake/ninja/ocl-icd
   but not xxd; adding vim-common (which ships xxd on RHEL-family
   distros) fixes the first build failure.

2. qrackbind_core.h: change bitLenInt alias uint16_t → uint8_t
   Qrack defines bitLenInt as a preprocessor macro (#define bitLenInt
   uint8_t when QBCAPPOW ≤ 7, the default).  qrackbind_core.h then
   does #undef bitLenInt / using bitLenInt = uint16_t.  This caused
   vector<uint16_t> (binding) to mismatch vector<unsigned char>/uint8_t
   (Qrack API — already macro-expanded to uint8_t during #include).
   The fix is to use uint8_t in the alias so both sides agree.

Result: qrackbind-0.1.0-cp312-cp312-manylinux_2_34_x86_64.whl produced
locally; smoke test passed (QrackSimulator(1).h(0) → 'smoke test OK').

fix: remove no-op build-* variants and dead cmake options

build-cpu, build-cuda, build-double, and build-no-simd all passed cmake
defines (ENABLE_OPENCL, ENABLE_CUDA, ENABLE_SSE3, ENABLE_AVX, FPPOW) that
were never used in CMakeLists.txt — Qrack's configuration is baked into the
pre-installed library and cannot be overridden at binding-compile time.

Remove the four justfile recipes and the five dead option() declarations.
Keep build-debug (cmake.build-type=Debug is a real scikit-build-core param),
build-verbose, build-define, cibuild, and cibuild-cpu (the latter correctly
controls Qrack's own cmake via CIBW_BEFORE_ALL).

update: Updated Phase 9 specification cibuildwheel info.

+++ bindings/qrackbind_core.h
@@ -17,10 +17,15 @@
 #include "common/pauli.hpp"
 #include "qcircuit.hpp"
 
-// Qrack defines these as preprocessor macros, we have to undefine them first
+// Qrack defines bitLenInt / bitCapInt as preprocessor macros.  Undefining them
+// gives us C++ type aliases we can refer to without macro-expansion hazards.
+// The alias types MUST match whatever Qrack's #define expanded to — otherwise
+// vector<bitLenInt> (our alias) ≠ vector<unsigned char> (Qrack's API), and the
+// compiler rejects implicit conversion.  Qrack's default QBCAPPOW=5 expands
+// both macros to uint8_t / uint64_t respectively, so we match those here.
 #undef bitLenInt
 #undef bitCapInt
-using bitLenInt = uint16_t;
+using bitLenInt = uint8_t;   // must match Qrack's #define bitLenInt (uint8_t when QBCAPPOW≤7)
 using bitCapInt = uint64_t;
 using real1_f   = float;
